+++ CHANGELOG.md
@@ -37,6 +37,8 @@ Delete any empty sections before publishing.
 
 - New icons: `trash`, `archive`, `share`, `link`, `tag`, `bookmark`, `expand-left`, `expand-up`, `expand-down`, `list-bullet`, `down-from-bracket`, `up-to-bracket`
 
+- New variant for tag component: `flat`
+
 ### Changed
 
 - Renamed icons for consistency: `checkmark` → `check`, `indeterminate` → `minus`, `logo` → `stack`, `download` → `down-to-bracket`, `upload` → `up-from-bracket`